+++ xivo_ctid_ng/plugins/conferences/bus_consume.py
@@ -16,6 +16,8 @@ class ConferencesBusEventHandler:
     def subscribe(self, bus_consumer):
         bus_consumer.on_ami_event('ConfbridgeJoin', self._notify_participant_joined)
         bus_consumer.on_ami_event('ConfbridgeLeave', self._notify_participant_left)
+        bus_consumer.on_ami_event('ConfbridgeMute', self._notify_participant_muted)
+        bus_consumer.on_ami_event('ConfbridgeUnmute', self._notify_participant_unmuted)
 
     def _notify_participant_joined(self, event):
         conference_id = int(event['Conference'])
@@ -52,3 +54,39 @@ class ConferencesBusEventHandler:
         participant = participant_schema.load(raw_participant).data
 
         self._notifier.participant_left(conference_id, participant)
+
+    def _notify_participant_muted(self, event):
+        conference_id = int(event['Conference'])
+        logger.debug('Participant in conference %s was muted', conference_id)
+        raw_participant = {
+            'id': event['Uniqueid'],
+            'caller_id_name': event['CallerIDName'],
+            'caller_id_num': event['CallerIDNum'],
+            'muted': True,
+            'answered_time': '0',
+            'admin': event['Admin'] == 'Yes',
+            'language': event['Language'],
+            'call_id': event['Uniqueid'],
+        }
+
+        participant = participant_schema.load(raw_participant).data
+
+        self._notifier.participant_muted(conference_id, participant)
+
+    def _notify_participant_unmuted(self, event):
+        conference_id = int(event['Conference'])
+        logger.debug('Participant in conference %s was unmuted', conference_id)
+        raw_participant = {
+            'id': event['Uniqueid'],
+            'caller_id_name': event['CallerIDName'],
+            'caller_id_num': event['CallerIDNum'],
+            'muted': False,
+            'answered_time': '0',
+            'admin': event['Admin'] == 'Yes',
+            'language': event['Language'],
+            'call_id': event['Uniqueid'],
+        }
+
+        participant = participant_schema.load(raw_participant).data
+
+        self._notifier.participant_unmuted(conference_id, participant)