+++ xivo_ctid_ng/plugins/conferences/resources.py
@@ -35,3 +35,21 @@ class ParticipantResource(AuthResource):
         tenant = Tenant.autodetect()
         self._service.kick_participant(tenant.uuid, conference_id, participant_id)
         return '', 204
+
+
+class ParticipantMuteResource(AuthResource):
+
+    def __init__(self, conferences_service):
+        self._service = conferences_service
+
+    @required_acl('ctid-ng.conferences.{conference_id}.participants.{participant_id}.mute.create')
+    def post(self, conference_id, participant_id):
+        tenant = Tenant.autodetect()
+        self._service.mute_participant(tenant.uuid, conference_id, participant_id)
+        return '', 204
+
+    @required_acl('ctid-ng.conferences.{conference_id}.participants.{participant_id}.mute.delete')
+    def delete(self, conference_id, participant_id):
+        tenant = Tenant.autodetect()
+        self._service.unmute_participant(tenant.uuid, conference_id, participant_id)
+        return '', 204